reformat(colo) rename folder `colorscheme` to `colo`
