+++ lua/colo/syntax.lua
